requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
soupsieve==2.6
urllib3==2.2.3
yarl==1.15.2
